Add stage-2 two-forward, ligand feedback, and bf16 dtype fix
- Configurable stage2 two-forward in configs/training.yaml
  (stage2_two_forward, stage2_ligand_weight, stage2_lowt_gamma)
- Two-forward implementation in molopt_score_model with
  apo-base differentiable interpolation and low-t weighting
- pl_model: optional unfreeze_prefixes for trunk unfreeze
  experiments; log loss_ligand2 alongside gate diagnostics
- Cross-attn gate index_copy: cast per-res to logits dtype to
  unblock bf16 autocast (fp32 bias would otherwise promote)
- validation/smoke_stage2.py and stage2_pocket_sensitivity.py
  for go/no-go diagnostics and single-GPU smoke tests

+++ models/ligand_residue_cross_attn.py
@@ -48,5 +48,7 @@ class LigandResidueCrossAttnGate(nn.Module):
             sim = q @ k.transpose(0, 1) * self.scale
             # for each residue, take the strongest ligand atom's attention logit
             per_res = sim.max(dim=0).values + self.bias
-            logits = logits.index_copy(0, res_idx, per_res)
+            # under bf16 autocast the fp32 bias promotes per_res to fp32; cast
+            # back to the logits buffer dtype so index_copy dtypes match.
+            logits = logits.index_copy(0, res_idx, per_res.to(logits.dtype))
         return logits
